Backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
supabase>=2.5.1
stripe>=10.1.0
python-dotenv>=1.0.1
